add password entry in registration page, delete password entry page
