+++ pages/registration/registration.tsx
@@ -6,6 +6,8 @@ import Link from "next/link";
 export default function Registration() {
     const [email, setEmail] = useState<string>('');
     const [login, setLogin] = useState<string>('');
+    const [password, setPassword] = useState<string>('');
+    const [passwordConfirmation, setPasswordConfirmation] = useState<string>('');
     const [checked, setChecked] = useState<boolean>(false);
     const [canSubmit, setCanSubmit] = useState<boolean>(false);
     const regExpEmail = /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
@@ -21,18 +23,33 @@ export default function Registration() {
 
     useEffect(() => {
         setCanSubmit(
-            validateEmail(email) && validateLogin(login) && checked
+            validateEmail(email) && validateLogin(login)
+            && checked && password.length >= 8 &&
+            password === passwordConfirmation
         );
     });
 
     const getLoginClassName = () => {
         if (validateLogin(login) || login === '') {
-           return '';
+            return '';
         } else return 'is-invalid'
     };
 
     const getEmailClassName = () => {
         if (validateEmail(email) || email === '') {
+            return '';
+        } else return 'is-invalid'
+    };
+
+    const getPasswordConfirmationClassName = () => {
+        if (passwordConfirmation === password || passwordConfirmation.length < 8
+            || password.length < 8 || passwordConfirmation === '') {
+            return '';
+        } else return 'is-invalid'
+    };
+
+    const getPasswordClassName = () => {
+        if (password.length >= 8 || password === '') {
             return '';
         } else return 'is-invalid'
     };
@@ -68,6 +85,30 @@ export default function Registration() {
                             onChange={(e) => setLogin(e.target.value)}
                         />
                         <Form.Label className='invalid-feedback'>Invalid Login</Form.Label>
+                    </Form.Group>
+
+                    <Form.Group className={styles.input}>
+                        <Form.Label className={styles.label}>Password</Form.Label>
+                        <Form.Control
+                            type='password'
+                            name='password'
+                            value={password}
+                            className={getPasswordClassName()}
+                            onChange={(e) => setPassword(e.target.value)}
+                        />
+                        <Form.Label className='invalid-feedback'>Password must be at least 8 symbols long</Form.Label>
+                    </Form.Group>
+
+                    <Form.Group className={styles.input}>
+                        <Form.Label className={styles.label}>Password confirmation</Form.Label>
+                        <Form.Control
+                            type='password'
+                            name='passwordConfirmation'
+                            value={passwordConfirmation}
+                            className={getPasswordConfirmationClassName()}
+                            onChange={(e) => setPasswordConfirmation(e.target.value)}
+                        />
+                        <Form.Label className='invalid-feedback'>Does not match password</Form.Label>
                     </Form.Group>
 
                     <Form.Group className='custom-control custom-checkbox'>
